+++ Userland/Libraries/CMakeLists.txt
@@ -26,6 +26,7 @@ add_subdirectory(LibGL)
 add_subdirectory(LibGPU)
 add_subdirectory(LibGUI)
 add_subdirectory(LibHTTP)
+add_subdirectory(LibIDL)
 add_subdirectory(LibIMAP)
 add_subdirectory(LibImageDecoderClient)
 add_subdirectory(LibIPC)
