+++ Userland/Libraries/LibIDL/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(SOURCES
+    IDLParser.cpp
+)
+
+serenity_lib(LibIDL idl)
+target_link_libraries(LibIDL LibCore)
